# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SimpleWatch API
  version: 0.1.0
...
paths:
...
  /api/v1/media/{mediaId}:
    get:
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: mediaId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  displayName:
                    type: string
                  state:
                    type: string
                    enum:
                      - scanning
                      - compatible
                      - incompatible
                      - failed
                      - published
                  bytes:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  mime:
                    nullable: true
                    type: string
                  compatibilityReasons:
                    type: array
                    items:
                      type: string
                  durationMs:
                    nullable: true
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  video:
                    type: object
                    properties:
                      codec:
                        nullable: true
                        type: string
                        enum:
                          - h264
                          - hevc
                      playbackSupport:
                        type: string
                        enum:
                          - broad
                          - device-dependent
                          - unsupported
                      width:
                        nullable: true
                        type: integer
                        exclusiveMinimum: true
                        maximum: 9007199254740991
                      height:
                        nullable: true
                        type: integer
                        exclusiveMinimum: true
                        maximum: 9007199254740991
                      fps:
                        nullable: true
                        type: number
                        minimum: 0
                      pixelFormat:
                        nullable: true
                        type: string
                    required:
                      - codec
                      - playbackSupport
                      - width
                      - height
                      - fps
                      - pixelFormat
                    additionalProperties: false
                  audio:
                    type: object
                    properties:
                      codec:
                        nullable: true
                        type: string
                      channels:
                        nullable: true
                        type: integer
                        exclusiveMinimum: true
                        maximum: 9007199254740991
                      sampleRate:
                        nullable: true
                        type: integer
                        exclusiveMinimum: true
                        maximum: 9007199254740991
                    required:
                      - codec
                      - channels
                      - sampleRate
                    additionalProperties: false
                  createdAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                  subtitles:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        language:
                          type: string
                        label:
                          type: string
                        format:
                          type: string
                          enum:
                            - webvtt
                      required:
                        - id
                        - language
                        - label
                        - format
                      additionalProperties: false
                required:
                  - id
                  - displayName
                  - state
                  - bytes
                  - mime
                  - compatibilityReasons
                  - durationMs
                  - video
                  - audio
                  - createdAt
                  - subtitles
                additionalProperties: false
  /api/v1/media/{mediaId}/content:
    get: &a6
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: mediaId
          required: true
      responses:
        "200":
          description: Default Response
    head: *a6
  /api/v1/uploads/authorize:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                  minLength: 1
                  maxLength: 255
                bytes:
                  type: integer
                  exclusiveMinimum: true
                  maximum: 53687091200
                mime:
                  type: string
                  minLength: 1
                  maxLength: 255
              required:
                - filename
                - bytes
                - mime
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploadId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  tusEndpoint:
                    type: string
                    format: uri
                  uploadToken:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                  maxChunkBytes:
                    type: integer
                    exclusiveMinimum: true
                    maximum: 9007199254740991
                required:
                  - uploadId
                  - tusEndpoint
                  - uploadToken
                  - expiresAt
                  - maxChunkBytes
                additionalProperties: false
  /api/v1/uploads/{uploadId}:
    get:
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: uploadId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  state:
                    type: string
                    enum:
                      - authorized
                      - uploading
                      - received
                      - scanning
                      - compatible
                      - incompatible
                      - failed
                      - published
                      - cancelled
                  filename:
                    type: string
                  mime:
                    type: string
                  declaredBytes:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  receivedBytes:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  expiresAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                required:
                  - id
                  - state
                  - filename
                  - mime
                  - declaredBytes
                  - receivedBytes
                  - expiresAt
                additionalProperties: false
    delete:
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: uploadId
          required: true
      responses:
        "204":
          description: Default Response
          content:
            application/json:
              schema:
                type: string
                nullable: true
                enum:
                  - null
  /api/v1/admin/media/{mediaId}/rescan:
    post:
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: mediaId
          required: true
      responses:
        "202":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                required:
                  - jobId
                additionalProperties: false
  /api/v1/admin/media/{mediaId}/subtitles:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                language:
                  type: string
                  minLength: 2
                  maxLength: 35
                label:
                  type: string
                  minLength: 1
                  maxLength: 80
                content:
                  type: string
                  minLength: 1
                  maxLength: 2097152
              required:
                - language
                - label
                - content
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: mediaId
          required: true
      responses:
        "202":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                required:
                  - jobId
                additionalProperties: false
  /api/v1/admin/media/{mediaId}:
    delete:
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: mediaId
          required: true
      responses:
        "204":
          description: Default Response
          content:
            application/json:
              schema:
                type: string
                nullable: true
                enum:
                  - null
  /api/v1/subtitles/{subtitleId}:
    get:
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: subtitleId
          required: true
      responses:
        "200":
          description: Default Response
  /api/v1/internal/uploads/complete:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uploadId:
                  type: string
                  format: uuid
                  pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                uploadToken:
                  type: string
                  minLength: 1
                filePath:
                  type: string
                  minLength: 1
                receivedBytes:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
              required:
                - uploadId
                - uploadToken
                - filePath
                - receivedBytes
      responses:
        "200":
          description: Default Response
  /api/v1/internal/tus/hooks:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
      responses:
        "200":
          description: Default Response
  /api/v1/internal/imports/sftp:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                  minLength: 1
                  maxLength: 255
                filePath:
                  type: string
                  minLength: 1
                bytes:
                  type: integer
                  exclusiveMinimum: true
                  maximum: 9007199254740991
              required:
                - filename
                - filePath
                - bytes
      responses:
        "200":
          description: Default Response
  /api/v1/internal/jobs/claim:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workerId:
                  type: string
                  minLength: 1
                  maxLength: 128
              required:
                - workerId
      responses:
        "200":
          description: Default Response
  /api/v1/internal/jobs/{jobId}/heartbeat:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                leaseToken:
                  type: string
                progress: {}
              required:
                - leaseToken
                - progress
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: jobId
          required: true
      responses:
        "200":
          description: Default Response
  /api/v1/internal/jobs/{jobId}/result:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                compatible:
                  type: boolean
                probe:
                  type: object
                  additionalProperties: {}
                reasons:
                  type: array
                  items:
                    type: string
                sha256:
                  type: string
                  pattern: ^[a-f0-9]{64}$
                bytes:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
                durationMs:
                  nullable: true
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
                finalPath:
                  type: string
                  minLength: 1
                leaseToken:
                  type: string
              required:
                - compatible
                - probe
                - reasons
                - sha256
                - bytes
                - durationMs
                - finalPath
                - leaseToken
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: jobId
          required: true
      responses:
        "200":
          description: Default Response
  /api/v1/internal/jobs/{jobId}/subtitle-result:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                leaseToken:
                  type: string
                finalPath:
                  type: string
                  minLength: 1
                sha256:
                  type: string
                  pattern: ^[a-f0-9]{64}$
                bytes:
                  type: integer
                  exclusiveMinimum: true
                  maximum: 2097152
              required:
                - leaseToken
                - finalPath
                - sha256
                - bytes
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: jobId
          required: true
      responses:
        "200":
          description: Default Response
...
servers:
  - url: https://watch.simplec.top